03.07 fix int size_t mixing

+++ KingdomRushLite/Scripts/Header/Manager/Concrete/WaveManager.h
@@ -12,20 +12,20 @@ class WaveManager :public Manager<WaveManager>
 	friend class Manager<WaveManager>;
 
 private:
-	std::vector<Wave> waveList;     //记录波次列表
-	int waveIdx = -1;               //当前正在进行的宏观波次索引
-	int eventIdx = 0;               //当前波次进行的微观生成事件的索引
+	std::vector<Wave> waveList;        //记录波次列表
+	int waveIdx = -1;                  //当前正在进行的宏观波次索引
+	int eventIdx = 0;                  //当前波次进行的微观生成事件的索引
 	
-	Timer newWaveTimer;             //用于唤起宏观波次的计时器
-	Timer newEventTimer;            //用于唤起微观生成事件的计时器
+	Timer newWaveTimer;                //用于唤起宏观波次的计时器
+	Timer newEventTimer;               //用于唤起微观生成事件的计时器
 
-	bool isWaveEnded = true;        //当前波次的所有生成事件执行完毕
+	bool isWaveEnded = true;           //当前波次的所有生成事件执行完毕
 
 public:
-	void OnUpdate(double);          //帧更新函数
+	void OnUpdate(double);             //帧更新函数
 
-	int GetCurrentWaveIdx() const;  //获取当前进行到的波次数
-	int GetTotalWaveNum() const;    //获取总波次数
+	size_t GetCurrentWaveIdx() const;  //获取当前进行到的波次数
+	size_t GetTotalWaveNum() const;    //获取总波次数
 
 private:
 	WaveManager();
